WorkAttendance/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
opencv-python==4.5.3.56
numpy==1.15.0
dlib==19.6.1
scipy==1.5.4
python-dotenv==0.19.0
requests==2.26.0